requirements.txt
discord.py>=2.2.0,<3.0.0
python-dotenv>=0.19.0
anthropic>=0.3.0
openai>=0.27.0
Pillow>=9.0.0
requests>=2.25.0
aiohttp>=3.7.0
replicate>=0.8.0